requirements.txt
scikit-learn
ramp
gdown
matplotlib
scipy
pandas
numpy
ramp-utils
ramp-workflow
scikit-image
torch
torchvision


